Unsubscribe기능 추가 / 테스트코드 작성 / 리팩토링 (다음엔 더 잘라서 pr..할게요!)

+++ OneCoin/Server/src/main/java/OneCoin/Server/chat/chatMessage/dto/ChatRequestDto.java
@@ -19,7 +19,7 @@ public class ChatRequestDto {
     @NotBlank
     private Long userId;
     @NotBlank
-    private Long chatRoomId;
+    private Integer chatRoomId;
     @NotBlank
     private String message;
